% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ir_spat}
\alias{ir_spat}
\title{Fits an spatial item response model using Stan
It requires
It requires
Missing values are not allowed in the response/covariates.}
\usage{
ir_spat(
  formula = formula,
  data = data,
  spat_model = "exp",
  itemtype = "3PLUS",
  abil = "user",
  diff = "id",
  y = y,
  coords = coords,
  iter = 4000,
  warmup = 2000,
  chains = 3,
  seed = seed,
  refresh = max(1000/100, 1),
  loglik = F
)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A long data frame containing the locations, dates, covariates and the response variable. It has to have the locID and date. No missing values are allowed in the covariates.}

\item{spat_model}{A string with the spatial correlation structure based on Euclidean distance models. The options are: exponential (exp), spherical (sph), gaussian (gau), car (Conditional Autoregressive). For no spatial autocorrelation use "none".}

\item{itemtype}{A string with the kind of item response model. Options: 3PLUS, 2PLUS, 1PLUS, 3PL, 2PL and 1PL. PLUS and PL models are spatial and non-spatial respectivelly.}

\item{abil}{A column indicating the user id}

\item{diff}{A column indicating the item id}

\item{y}{binary response variable (1 = correct, 0 = otherwise).}

\item{coords}{A vector with the lon and lat for Euclidean distance models.}

\item{iter}{Number of iterations}

\item{warmup}{Warm up samples}

\item{chains}{Number of chains}

\item{seed}{(optional) A seed for reproducibility}

\item{refresh}{The refresh rate showing the progress}

\item{loglik}{If the log-likelihood will be computed}
}
\value{
A list with the model fit
}
\description{
Fits an spatial item response model using Stan
It requires
It requires
Missing values are not allowed in the response/covariates.
}
\details{
Missing values are not allowed in the covariates and they must be imputed before using ir_spat(). Many options can be found in https://cran.r-project.org/web/views/MissingData.html
}
\examples{
# ir_spat(formula = site ~ -1 + Standing + Resting + Moving + Eating + Interacting + Babies,
#                    data = s,
#                    spat_model = 'car',
#                    itemtype = '3PLUS',
#                    iter = 40,
#                    warmup = 20,
#                    chains = 3,
#                    refresh = 100
#                    )
}
