% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mungeCARdata4stan}
\alias{mungeCARdata4stan}
\title{A function that finds the neighbours of each point in the vornoi graph.}
\usage{
mungeCARdata4stan(adjBUGS, numBUGS)
}
\arguments{
\item{adjBUGS}{a list of region ids for the adjacent regions}

\item{numBUGS}{a list of the number of neighbours for each region}
}
\value{
a list containing parameters required by the stan model.
}
\description{
A function that finds the neighbours of each point in the vornoi graph.
}
\examples{
nb2 <- voronoi(matrix(c(data$lon, data$lat), ncol=2))

#https://gis.stackexchange.com/questions/237810/adjacency-matrix-not-including-vertices
polyids = seq_along(nb2)
adjMat = matrix(FALSE, ncol = length(nb2), nrow = length(nb2))

for (ii in polyids) {
  for (jj in setdiff(polyids, seq_len(ii))) {
    adjMat[ii, jj] =
      ifelse(class(gIntersection(nb2[ii, ], nb2[jj, ])) == 'SpatialLines',1,0)
  }
}

#use symmetry for the other half
adjMat[lower.tri(adjMat)] = t(adjMat)[lower.tri(adjMat)]

num <- apply(W,1,sum)
adj <- c(unlist(apply(W,1,function(x) which(x == 1) ) ))
munged_data <- mungeCARdata4stan(adjBUGS = adj,numBUGS = num)
}
\author{
Mitzi Morris
}
