% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staircase.R
\docType{package}
\name{staircase-package}
\alias{staircase}
\alias{staircase-package}
\title{staircase: (STA)tistical (I)tem (R)esponse models for (C)itizen (S)cience in (E)cology}
\description{
The package fits item response models to citizen science ecological data. It accounts for spatial dependence in the difficulties. The functions in this packages are utilised in this paper:\cr\cr Santos-Fernández, Edgar & Mengersen, Kerrie. (2020). Bayesian item response models for citizen science ecological data. 10.13140/RG.2.2.35569.63847.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EdgarSantos-Fernandez/staircase,}
}

}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Edgar edgar.santosfernandez@qut.edu.au Santos-Fernández \email{aut}
}

}
\keyword{internal}
